+++ examples/caretta-id-cli/src/command/random.rs
@@ -1,0 +1,12 @@
+use caretta_id::CarettaId;
+use clap::Args;
+
+#[derive(Args, Debug)]
+/// Generate random caretta-id
+pub struct RandomCommandArgs;
+
+impl RandomCommandArgs {
+    pub fn run(self) {
+        println!("{}", rand::random::<CarettaId>())
+    }
+}