+++ examples/caretta-id-cli/tests/timestamp.rs
@@ -1,0 +1,12 @@
+use caretta_id::{CarettaId, CarettaIdD, CarettaIdQ, CarettaIdS, CarettaIdT};
+use std::{path::PathBuf, process::Command};
+
+#[test]
+fn now_unix() {
+    let path = PathBuf::from(std::env!("CARGO_BIN_EXE_caretta-id-cli"));
+    let output =
+        String::from_utf8_lossy(&Command::new(path).arg("timestamp").arg("--unix").output().unwrap().stdout)
+            .trim()
+            .to_owned();
+    let _ = output.parse::<CarettaId>().unwrap();
+}